+++ internal/mwdd/cmd.go
@@ -252,8 +252,9 @@ func NewServiceExposeCmdP(name *string) *cobra.Command {
 		Use:   "expose",
 		Short: "Expose a port in a running container",
 		Example: heredoc.Doc(`
-		expose --external-port 8899
-		expose --external-port 8899 --internal-port 80
+		expose
+		expose --external-port 1234
+		expose --external-port 1234 --internal-port 80
 		`),
 		Run: func(cmd *cobra.Command, args []string) {
 			dereferencedName := *name
@@ -265,6 +266,11 @@ func NewServiceExposeCmdP(name *string) *cobra.Command {
 			cli := docker.NewClientFromEnvOrPanic()
 			containerID, err := m.DockerCompose().ContainerID(dereferencedName)
 			if err != nil {
+				// unable to execute command, no container found for service: mysql
+				if strings.Contains(err.Error(), "no container found") {
+					logrus.Error("Container must be running before you can expose a port")
+					return
+				}
 				panic(err)
 			}
 
@@ -277,7 +283,7 @@ func NewServiceExposeCmdP(name *string) *cobra.Command {
 					panic(err)
 				}
 
-				// Get the DEFAULT_EXPOSE_PORT environemtn variable from the containerJson if set
+				// Get the DEFAULT_EXPOSE_PORT environment variable from the containerJson if set
 				for _, env := range containerJson.Config.Env {
 					if strings.HasPrefix(env, "DEFAULT_EXPOSE_PORT=") {
 						internalPort = strings.Split(env, "=")[1]
@@ -383,7 +389,7 @@ func NewImageCmdP(service *string) *cobra.Command {
 
 	set := &cobra.Command{
 		Use:   "set",
-		Short: "Sets the image to use for the service as an overrides environemtn variable",
+		Short: "Sets the image to use for the service as an overrides environment variable",
 		Run: func(cmd *cobra.Command, args []string) {
 			dereferencedService := *service
 			DefaultForUser().EnsureReady()